BACKLOG-19917: Dispatch actions from picker dialog
